+++ commands/Core/ping.js
@@ -1,0 +1,8 @@
+module.exports = {
+	name: 'ping',
+	description: 'Ping!',
+	cooldown: 5,
+	execute(message) {
+		message.channel.send('Pong!');
+	},
+};
