+++ commands/Core/restart.js
