--- commands/avatar.js
@@ -1,16 +1,0 @@
-module.exports = {
-	name: 'avatar',
-	description: 'Get the avatar URL of the tagged user(s), or your own avatar..',
-	aliases: ['icon', 'pfp'],
-	execute(message) {
-		if (!message.mentions.users.size) {
-			return message.channel.send(`Your avatar: ${message.author.displayAvatarURL}`);
-		}
-
-		const avatarList = message.mentions.users.map(user => {
-			return `${user.username}'s avatar: ${user.displayAvatarURL}`;
-		});
-
-		message.channel.send(avatarList);
-	},
-};
