--- commands/ping.js
@@ -1,8 +1,0 @@
-module.exports = {
-	name: 'ping',
-	description: 'Ping!',
-	cooldown: 5,
-	execute(message) {
-		message.channel.send('Pong!');
-	},
-};
