--- commands/user-info.js
@@ -1,7 +1,0 @@
-module.exports = {
-	name: 'user-info',
-	description: 'Display info about yourself.',
-	execute(message) {
-		message.channel.send(`Your username: ${message.author.username}\nYour ID: ${message.author.id}`);
-	},
-};
